Download model file from S3 on websocket open
Downloading the file every time this happens is slow and unnecessary,
but it keeps things simple for now.
Eventually, the plan is to download the file and load it into a a Keras
only when we either haven't downloaded the model before, or the version on
S3 has been updated more recently than what we downloaded before.